daemon: add blockdag provider, add comments
