+++ xelis_daemon/src/core/storage/providers/mod.rs
@@ -8,6 +8,7 @@ mod balance;
 mod client_protocol;
 mod transaction;
 mod block;
+mod blockdag;
 
 pub use asset::AssetProvider;
 pub use blocks_at_height::BlocksAtHeightProvider;
@@ -19,3 +20,4 @@ pub use balance::BalanceProvider;
 pub use client_protocol::ClientProtocolProvider;
 pub use transaction::TransactionProvider;
 pub use block::BlockProvider;
+pub use blockdag::BlockDagProvider;
